Add Deflate Compression to existing code.
* Server now is capable of compressing response
with the help of zlib to generate message with
the header of "Content-Encoding" if the request
message contains the request header "Accept-
Encoding: deflate". This can dramatically reduce
the body of response message to the degree of
about 50%, which is out of my expectation.

* The server does not support "PUT" and "HEAD"
request method, so we should indicate it by
informing the client with the help of HTTP 501
status code "Not Implemented".